+++ lib/msal-browser/src/interaction_client/PopupClient.ts
@@ -227,15 +227,16 @@ export class PopupClient extends StandardInteractionClient {
                 account: validRequest.account,
             });
 
-            const isNativeBroker = NativeMessageHandler.isNativeAvailable(
-                this.config,
-                this.logger,
-                this.nativeMessageHandler,
-                request.authenticationScheme
-            );
+            const isPlatformBroker =
+                NativeMessageHandler.isPlatformBrokerAvailable(
+                    this.config,
+                    this.logger,
+                    this.nativeMessageHandler,
+                    request.authenticationScheme
+                );
             // Start measurement for server calls with native brokering enabled
             let fetchNativeAccountIdMeasurement;
-            if (isNativeBroker) {
+            if (isPlatformBroker) {
                 fetchNativeAccountIdMeasurement =
                     this.performanceClient.startMeasurement(
                         PerformanceEvents.FetchAccountIdWithNativeBroker,
@@ -246,7 +247,7 @@ export class PopupClient extends StandardInteractionClient {
             // Create acquire token url.
             const navigateUrl = await authClient.getAuthCodeUrl({
                 ...validRequest,
-                nativeBroker: isNativeBroker,
+                platformBroker: isPlatformBroker,
             });
 
             // Create popup interaction handler.