+++ lib/msal-browser/src/packageMetadata.ts
@@ -1,3 +1,3 @@
 /* eslint-disable header/header */
 export const name = "@azure/msal-browser";
-export const version = "3.28.0";
+export const version = "4.0.2";