+++ lib/msal-node/test/cache/Storage.spec.ts
@@ -141,7 +141,7 @@ describe("Storage tests for msal-node: ", () => {
         expect(account).toEqual(fetchedAccount);
     });
 
-    it("setAccount() and getAccount() tests", () => {
+    it("setAccount() and getAccount() tests", async () => {
         const nodeStorage = new NodeStorage(
             logger,
             clientId,
@@ -180,63 +180,10 @@ describe("Storage tests for msal-node: ", () => {
             mockAccountData
         );
         expect(mockAccountEntity).toBeInstanceOf(AccountEntity);
-        nodeStorage.setAccount(mockAccountEntity);
+        await nodeStorage.setAccount(mockAccountEntity);
         expect(
             nodeStorage.getAccount(mockAccountEntity.generateAccountKey())
         ).toEqual(mockAccountEntity);
-    });
-
-    it("getAccount() updates an outdated (single-tenant) account cache entry", () => {
-        const nodeStorage = new NodeStorage(
-            logger,
-            clientId,
-            DEFAULT_CRYPTO_IMPLEMENTATION
-        );
-        nodeStorage.setInMemoryCache(inMemoryCache);
-        const outdatedAccountKey = "uid.utid3-login.microsoftonline.com-utid3";
-
-        const outdatedAccountData = {
-            username: "janedoe@microsoft.com",
-            name: "Jane Doe",
-            localAccountId: "uid",
-            realm: "utid3",
-            environment: "login.microsoftonline.com",
-            homeAccountId: "uid.utid3",
-            authorityType: "MSSTS",
-            clientInfo: "eyJ1aWQiOiJ1aWQxIiwgInV0aWQiOiJ1dGlkMSJ9",
-        };
-
-        let outdatedMockAccountEntity = CacheManager.toObject(
-            new AccountEntity(),
-            outdatedAccountData
-        );
-
-        let updatedMockAccountEntity = CacheManager.toObject(
-            new AccountEntity(),
-            {
-                ...outdatedAccountData,
-                tenantProfiles: [
-                    {
-                        tenantId: "utid3",
-                        localAccountId: "uid",
-                        name: "Jane Doe",
-                        isHomeTenant: true,
-                    },
-                ],
-            }
-        );
-        const updatedAccountKey = updatedMockAccountEntity.generateAccountKey();
-        expect(outdatedMockAccountEntity).toBeInstanceOf(AccountEntity);
-        // Set an outdated account
-        nodeStorage.setAccount(outdatedMockAccountEntity);
-        expect(nodeStorage.getItem(outdatedAccountKey)).toEqual(
-            outdatedAccountData
-        );
-
-        // Get account should update and return updated account
-        expect(nodeStorage.getAccount(updatedAccountKey)).toEqual(
-            updatedMockAccountEntity
-        );
     });
 
     it("setCache() and getCache() tests - tests for an accessToken", () => {
@@ -270,7 +217,7 @@ describe("Storage tests for msal-node: ", () => {
         expect(readCache[accessTokenKey]).toEqual(accessToken);
     });
 
-    it("setAccessTokenCredential() and getAccessTokenCredential() tests", () => {
+    it("setAccessTokenCredential() and getAccessTokenCredential() tests", async () => {
         const nodeStorage = new NodeStorage(
             logger,
             clientId,
@@ -294,7 +241,7 @@ describe("Storage tests for msal-node: ", () => {
             extendedExpiresOn: "4600",
         };
 
-        nodeStorage.setAccessTokenCredential(accessToken);
+        await nodeStorage.setAccessTokenCredential(accessToken);
         const fetchedAccessToken =
             nodeStorage.getAccessTokenCredential(accessTokenKey);
         const invalidAccessToken = nodeStorage.getAccessTokenCredential(
@@ -305,7 +252,7 @@ describe("Storage tests for msal-node: ", () => {
         expect(invalidAccessToken).toBeNull();
     });
 
-    it("setIdTokenCredential() and getIdTokenCredential() tests", () => {
+    it("setIdTokenCredential() and getIdTokenCredential() tests", async () => {
         const nodeStorage = new NodeStorage(
             logger,
             clientId,
@@ -325,7 +272,7 @@ describe("Storage tests for msal-node: ", () => {
             realm: "samplerealm",
         };
 
-        nodeStorage.setIdTokenCredential(idToken);
+        await nodeStorage.setIdTokenCredential(idToken);
 
         const fetchedIdToken = nodeStorage.getIdTokenCredential(idTokenKey);
         const invalidIdToken =
@@ -335,7 +282,7 @@ describe("Storage tests for msal-node: ", () => {
         expect(invalidIdToken).toBeNull();
     });
 
-    it("setRefreshTokenCredential() and getRefreshTokenCredential() tests", () => {
+    it("setRefreshTokenCredential() and getRefreshTokenCredential() tests", async () => {
         const nodeStorage = new NodeStorage(
             logger,
             clientId,
@@ -355,7 +302,7 @@ describe("Storage tests for msal-node: ", () => {
             realm: "samplerealm",
         };
 
-        nodeStorage.setRefreshTokenCredential(refreshToken);
+        await nodeStorage.setRefreshTokenCredential(refreshToken);
 
         const fetchedRefreshToken =
             nodeStorage.getRefreshTokenCredential(refreshTokenKey);