+++ lib/msal-node/test/client/ManagedIdentitySources/Imds.spec.ts
@@ -18,7 +18,6 @@ import {
     THREE_SECONDS_IN_MILLI,
     getCacheKey,
 } from "../../test_kit/StringConstants.js";
-
 import {
     ManagedIdentityNetworkClient,
     ManagedIdentityNetworkErrorClient,
@@ -26,6 +25,7 @@ import {
     userAssignedClientIdConfig,
     managedIdentityRequestParams,
     systemAssignedConfig,
+    userAssignedResourceIdConfig,
 } from "../../test_kit/ManagedIdentityTestUtils.js";
 import {
     DEFAULT_MANAGED_IDENTITY_ID,
@@ -48,13 +48,12 @@ import {
     createManagedIdentityError,
 } from "../../../src/error/ManagedIdentityError.js";
 import { mockCrypto } from "../ClientTestUtils.js";
-import {
-    CacheKVStore,
-    ClientCredentialClient,
-    NodeStorage,
-} from "../../../src/index.js";
 // NodeJS 16+ provides a built-in version of setTimeout that is promise-based
 import { setTimeout } from "timers/promises";
+import { ClientCredentialClient } from "../../../src/client/ClientCredentialClient.js";
+import { NodeStorage } from "../../../src/cache/NodeStorage.js";
+import { CacheKVStore } from "../../../src/cache/serializer/SerializerTypes.js";
+import { ManagedIdentityUserAssignedIdQueryParameterNames } from "../../../src/client/ManagedIdentitySources/BaseManagedIdentitySource.js";
 
 describe("Acquires a token successfully via an IMDS Managed Identity", () => {
     // IMDS doesn't need environment variables because there is a default IMDS endpoint
@@ -79,14 +78,6 @@ describe("Acquires a token successfully via an IMDS Managed Identity", () => {
         },
         managedIdentityIdParams: {
             userAssignedObjectId: MANAGED_IDENTITY_RESOURCE_ID,
-        },
-    };
-    const userAssignedResourceIdConfig: ManagedIdentityConfiguration = {
-        system: {
-            networkClient,
-        },
-        managedIdentityIdParams: {
-            userAssignedResourceId: MANAGED_IDENTITY_RESOURCE_ID,
         },
     };
 
@@ -126,6 +117,11 @@ describe("Acquires a token successfully via an IMDS Managed Identity", () => {
         });
 
         test("acquires a User Assigned Resource Id token", async () => {
+            const sendGetRequestAsyncSpy: jest.SpyInstance = jest.spyOn(
+                networkClient,
+                <any>"sendGetRequestAsync"
+            );
+
             const managedIdentityApplication: ManagedIdentityApplication =
                 new ManagedIdentityApplication(userAssignedResourceIdConfig);
             expect(managedIdentityApplication.getManagedIdentitySource()).toBe(
@@ -140,6 +136,22 @@ describe("Acquires a token successfully via an IMDS Managed Identity", () => {
             expect(networkManagedIdentityResult.accessToken).toEqual(
                 DEFAULT_USER_SYSTEM_ASSIGNED_MANAGED_IDENTITY_AUTHENTICATION_RESULT.accessToken
             );
+
+            const url: URLSearchParams = new URLSearchParams(
+                sendGetRequestAsyncSpy.mock.lastCall[0]
+            );
+            expect(
+                url.has(
+                    ManagedIdentityUserAssignedIdQueryParameterNames.MANAGED_IDENTITY_RESOURCE_ID_IMDS
+                )
+            ).toBe(true);
+            expect(
+                url.get(
+                    ManagedIdentityUserAssignedIdQueryParameterNames.MANAGED_IDENTITY_RESOURCE_ID_IMDS
+                )
+            ).toEqual(MANAGED_IDENTITY_RESOURCE_ID);
+
+            jest.restoreAllMocks();
         });
     });