+++ samples/msal-browser-samples/VanillaJSTestApp2.0/app/client-capabilities/test/browser.spec.ts
@@ -48,7 +48,7 @@ describe("Browser tests", function () {
     let page: puppeteer.Page;
     let BrowserCache: BrowserCacheUtils;
     beforeEach(async () => {
-        context = await browser.createIncognitoBrowserContext();
+        context = await browser.createBrowserContext();
         page = await context.newPage();
         page.setDefaultTimeout(ONE_SECOND_IN_MS * 5);
         BrowserCache = new BrowserCacheUtils(page, "sessionStorage");
@@ -98,9 +98,7 @@ describe("Browser tests", function () {
         expect(tokenStore.idTokens).toHaveLength(1);
         expect(tokenStore.accessTokens).toHaveLength(1);
         expect(tokenStore.refreshTokens).toHaveLength(1);
-        const cachedAccount = await BrowserCache.getAccountFromCache(
-            tokenStore.idTokens[0]
-        );
+        const cachedAccount = await BrowserCache.getAccountFromCache();
         const defaultCachedToken =
             await BrowserCache.accessTokenForScopesExists(
                 tokenStore.accessTokens,