+++ samples/msal-node-samples/auth-code/test/auth-code-b2c-msa.spec.ts
@@ -104,7 +104,7 @@ describe("Auth Code B2C Tests (msa account)", () => {
         });
 
         beforeEach(async () => {
-            context = await browser.createIncognitoBrowserContext();
+            context = await browser.createBrowserContext();
             page = await context.newPage();
             page.setDefaultTimeout(5000);
             page.on("dialog", async (dialog) => {