+++ samples/msal-react-samples/b2c-sample/test/msa-account.spec.ts
@@ -50,7 +50,7 @@ describe("B2C user-flow tests (msa account)", () => {
     });
 
     beforeEach(async () => {
-        context = await browser.createIncognitoBrowserContext();
+        context = await browser.createBrowserContext();
         page = await context.newPage();
         page.setDefaultTimeout(5000);
         BrowserCache = new BrowserCacheUtils(page, "localStorage");
@@ -89,7 +89,7 @@ describe("B2C user-flow tests (msa account)", () => {
         );
 
         // Verify UI now displays logged in content
-        await page.waitForXPath("//header[contains(., 'Welcome,')]");
+        await page.waitForSelector("xpath/.//header[contains(., 'Welcome,')]");
         await screenshot.takeScreenshot(page, "Signed in with the policy");
 
         // Verify tokens are in cache
@@ -98,9 +98,7 @@ describe("B2C user-flow tests (msa account)", () => {
         expect(tokenStoreBeforeEdit.accessTokens.length).toBe(1);
         expect(tokenStoreBeforeEdit.refreshTokens.length).toBe(1);
         expect(
-            await BrowserCache.getAccountFromCache(
-                tokenStoreBeforeEdit.idTokens[0]
-            )
+            await BrowserCache.getAccountFromCache()
         ).not.toBeNull();
         expect(
             await BrowserCache.accessTokenForScopesExists(
@@ -127,8 +125,8 @@ describe("B2C user-flow tests (msa account)", () => {
                 `window.location.href.startsWith("http://localhost:${port}")`
             ),
             page.waitForSelector("#idTokenClaims"),
-            page.waitForXPath(
-                "//*[@id=\"interactionStatus\"]/center[contains(., 'ssoSilent success')]",
+            page.waitForSelector(
+                "::-p-xpath(//*[@id=\"interactionStatus\"]/center[contains(., 'update success')])",
                 { timeout: 4000 }
             ),
         ]);
@@ -145,14 +143,10 @@ describe("B2C user-flow tests (msa account)", () => {
         expect(tokenStoreAfterEdit.accessTokens.length).toBe(1);
         expect(tokenStoreAfterEdit.refreshTokens.length).toBe(2); // 1 for each policy
         expect(
-            await BrowserCache.getAccountFromCache(
-                tokenStoreAfterEdit.idTokens[0]
-            )
+            await BrowserCache.getAccountFromCache()
         ).not.toBeNull();
         expect(
-            await BrowserCache.getAccountFromCache(
-                tokenStoreAfterEdit.idTokens[1]
-            )
+            await BrowserCache.getAccountFromCache()
         ).not.toBeNull(); // new account after edit
         expect(
             await BrowserCache.accessTokenForScopesExists(