+++ lib/msal-node/CHANGELOG.md
@@ -1,8 +1,34 @@
 # Change Log - @azure/msal-node
 
-This log was last generated on Tue, 19 Nov 2024 17:21:57 GMT and should not be manually modified.
+<!-- This log was last generated on Wed, 22 Jan 2025 21:45:31 GMT and should not be manually modified. -->
 
 <!-- Start content -->
+
+## 3.1.0
+
+Wed, 22 Jan 2025 21:45:31 GMT
+
+### Minor changes
+
+- Added Machine Learning as a Managed Identity Source #7512 (rginsburg@microsoft.com)
+- Bump @azure/msal-common to v15.0.2
+- Bump eslint-config-msal to v0.0.0
+- Bump rollup-msal to v0.0.0
+
+### Patches
+
+- Fixed incorrect IMDS resource ID query parameter (rginsburg@microsoft.com)
+- Avoid calling class methods in their own constructors (thomas.norling@microsoft.com)
+
+## 3.0.1
+
+Wed, 15 Jan 2025 05:05:17 GMT
+
+### Breaking Changes
+
+- NodeStorage is no longer exported
+
+Please see the [migration guide](https://github.com/AzureAD/microsoft-authentication-library-for-js/blob/dev/lib/msal-node/docs/v2-migration.md) for more details.
 
 ## 2.16.2